Moves the getCharacterToPrint function to a new module called simple
